+++ Examples/Resources/SqlRS/1-DefaultConfiguration.ps1
@@ -1,11 +1,11 @@
 <#
 .EXAMPLE
-    This example performs a default SSRS configuration. It will initialize SSRS
-    and register default Report Server Web Service and Report Manager URLs:
+    This example performs a default SQL Server Reporting Services configuration.
+    It will initialize SQL Server Reporting Services and register default
+    Report Server Web Service and Report Manager URLs.
 
-    http://localhost:80/ReportServer (Report Server Web Service)
-
-    http://localhost:80/Reports (Report Manager)
+    Report Server Web Service: http://localhost:80/ReportServer
+    Report Manager: http://localhost:80/Reports
 #>
 Configuration Example
 {